requirements.txt
numpy>=1.21.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
langchain>=0.0.300
openai>=1.0.0
anthropic>=0.5.0
fastapi>=0.100.0
uvicorn>=0.23.0
chromadb>=0.4.0
faiss-cpu>=1.7.4
pydantic>=2.0.0
graphene>=3.3.0
